% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainFunctions.R
\name{bearing_to_latlon}
\alias{bearing_to_latlon}
\title{bearing_to_latlon}
\usage{
bearing_to_latlon(origin_lat, origin_lon, bearing, gc_dist)
}
\arguments{
\item{origin_lat}{The origin latitude}

\item{origin_lon}{The origin longitude}

\item{bearing}{The angle in degrees relative to due north}

\item{gc_dist}{The greater circle distance in kilometres}
}
\description{
Calculate destination lat/lon given an origin, a bearing and a greater circle distance of travel
}
\examples{
origin_latLon <- c(0,0)
some_bearing <- 50
some_gc_dist <- 27
bearing_to_latlon(origin_lat = origin_latLon[1], origin_lon = origin_latLon[2], bearing = some_bearing, gc_dist = some_gc_dist)
}
