% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{bind_data}
\alias{bind_data}
\title{Bind data to project}
\usage{
bind_data(project, df, data_type, name = NULL,
  check_delete_output = TRUE)
}
\arguments{
\item{project}{an \code{rgeoprofile_project}, as produced by the function
\code{rgeoprofile_project()}}

\item{df}{a dataframe with columns that must conform to the following rules:
\itemize{
  \item for \code{data_type = "counts"}, data must have columns
  "longitude", "latitude" and "counts".
  \item for \code{data_type = "prevalence"}, data must have columns
  "longitude", "latitude", "tested" and "positive"
  }}

\item{data_type}{the type of data, either "counts" or "prevalence".}

\item{name}{optional name of the data set to aid in record keeping}

\item{check_delete_output}{whether to prompt the user before overwriting
existing data}
}
\description{
Load data into a \code{rgeoprofile_project} prior to analysis.
  Data must be formatted as a dataframe with samples in rows and loci in
  columns. If individuals are polyploid then multiple rows can be used per
  sample. Ploidy is allowed to vary between samples, and can be specified in
  multiple ways.
}
