% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{cartesian_to_latlon}
\alias{cartesian_to_latlon}
\title{cartesian_to_latlon}
\usage{
cartesian_to_latlon(centre_lat, centre_lon, data_x, data_y)
}
\arguments{
\item{centre_lat}{The centre latitude}

\item{centre_lon}{The centre longitude}

\item{data_x}{Cartesian x co-ordinate}

\item{data_y}{Cartesian y co-ordinate}
}
\description{
Transform Cartesian Co-ordinates into Latitude and Longitude
  points assuming a Lat_Long centre
}
\examples{
centre <- c(1,5)
data <- c(2,3)
cartesian_to_latlon(centre_lat = centre[1], centre_lon = centre[2],
                    data_x = data[1], data_y = data[2])
}
