% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{check_silverblaze_loaded}
\alias{check_silverblaze_loaded}
\title{Check that silverblaze package has loaded successfully}
\usage{
check_silverblaze_loaded()
}
\description{
Simple function to check that silverblaze package has loaded 
  successfully. Prints "silverblaze loaded successfully!" if so.
}
