% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{dist_gc}
\alias{dist_gc}
\title{Pairwise Great-Circle Distance}
\usage{
dist_gc(points)
}
\arguments{
\item{points}{a matrix with two columns specifying latitude and longitude}
}
\description{
Analogue of \code{dist()} function, but returning great-circle
  distances.
}
\examples{
some_points <- cbind(1:10, 1:10)
dist_gc(some_points)
}
