% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{get_ESS}
\alias{get_ESS}
\title{Get ESS}
\usage{
get_ESS(project, K = NULL)
}
\arguments{
\item{project}{an RgeoProfile project, as produced by the function
\code{rgeoprofile_project()}}

\item{K}{get ESS for this value of K}
}
\description{
Returns effective sample size (ESS) of chosen model run.
}
\examples{
\dontshow{p <- rgeoprofile_file("tutorial1_project.rds")}
get_ESS(project = p, K = 1)
get_ESS(project = p, K = 2)
get_ESS(project = p, K = 3)
}
