% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{get_hitscores}
\alias{get_hitscores}
\title{Get hitscores}
\usage{
get_hitscores(project, source_lon, source_lat, ring_search = TRUE)
}
\arguments{
\item{project}{an RgeoProfile project, as produced by the function
\code{rgeoprofile_project()}}

\item{source_lon}{longitudes of known sources}

\item{source_lat}{latitudes of known sources}

\item{ring_search}{Option to compute ring search hitscores}
}
\description{
Get hitscores
}
