% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{latlon_to_bearing}
\alias{latlon_to_bearing}
\title{latlon_to_bearing}
\usage{
latlon_to_bearing(origin_lat, origin_lon, dest_lat, dest_lon)
}
\arguments{
\item{origin_lat}{The origin latitude}

\item{origin_lon}{The origin longitude}

\item{dest_lat}{The destination latitude}

\item{dest_lon}{The destination longitude}
}
\description{
Calculate distance between latitude/longitude points
}
\examples{
origin_latLon <- c(0,0)
dest_latLon <- c(1,1)
latlon_to_bearing(origin_lat = origin_latLon[1], origin_lon = origin_latLon[2],
dest_lat = dest_latLon[1], dest_lon = dest_latLon[2])
}
