% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{latlon_to_cartesian}
\alias{latlon_to_cartesian}
\title{latlon_to_cartesian}
\usage{
latlon_to_cartesian(centre_lat, centre_lon, data_lat, data_lon)
}
\arguments{
\item{centre_lat}{The centre latitude}

\item{centre_lon}{The centre longitude}

\item{data_lat}{The data latitude}

\item{data_lon}{The data longitude}
}
\description{
Convert from latitude/longitude to cartesian co-ordinates
}
\examples{
centre_latLon <- c(0,1)
data_latLon <- c(0,1)
latlon_to_cartesian(centre_lat = centre_latLon[1], centre_lon = centre_latLon[2],
data_lat = data_latLon[1], data_lon = data_latLon[2])
}
