% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{logSum}
\alias{logSum}
\title{logSum}
\usage{
logSum(logx)
}
\arguments{
\item{logx}{A vector of values in log space}
}
\description{
Sum a vector of values together in log space, while avoiding
  underflow issues. Takes log(x) values as input and returns log(sum(x)) as
  output.
}
\examples{
x <- log(seq(5e-49, 5e-48, 1e-49))
exp(sum(x))
exp(logSum(x))
}
