% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{new_set}
\alias{new_set}
\title{Create new parameter set}
\usage{
new_set(project, spatial_prior = NULL, name = "(no name)",
  sigma_model = "single", dispersal_model = "normal",
  sigma_prior_mean = 1, sigma_prior_sd = 1,
  expected_popsize_model = "single",
  expected_popsize_prior_mean = 1000, expected_popsize_prior_sd = 20,
  sentinel_radius = 0.2, n_binom = FALSE, alpha_prior_mean = 1,
  alpha_prior_sd = 100)
}
\arguments{
\item{project}{an rgeoprofile_project, as produced by the function
\code{rgeoprofile_project()}.}

\item{spatial_prior}{a raster file defining the spatial prior. Precision
values are taken from this raster if it is defined.}

\item{name}{an optional name for the parameter set.}

\item{sigma_model}{set as \code{"single"} to assume the same dispersal
distance for all sources, or \code{"independent"} to assume an
independently drawn dispersal distance for each source.}

\item{dispersal_model}{distribute points via a "normal", "cauchy" or 
"laplace" model}

\item{sigma_prior_mean}{the prior mean of the parameter sigma (km).}

\item{sigma_prior_sd}{the prior standard deviation of the parameter sigma
(km). Set to 0 to use a fixed value for sigma (fixed at
\code{sigma_prior_mean}).}

\item{expected_popsize_model}{set as \code{"single"} to assume the same
number of events for all sources, or \code{"independent"} to assume an
independently drawn number of events for each source.}

\item{expected_popsize_prior_mean}{the prior mean of the expected total
population size.}

\item{expected_popsize_prior_sd}{the prior standard deviation of the expected
total population size. Set to 0 to use a fixed value (fixed at
\code{expected_popsize_prior_mean}).}

\item{sentinel_radius}{the observation radius of sentinel sites.}

\item{n_binom}{set to true or false, decide if a negative binomial model should
be run for a set of count data.}

\item{alpha_prior_mean}{the prior mean alpha.}

\item{alpha_prior_sd}{the prior standard deviation of alpha.}
}
\description{
Create a new parameter set within an \code{rgeoprofile_project}. The new
  parameter set becomes the active set once created.
}
