% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{optimise_beta}
\alias{optimise_beta}
\title{Optimise beta values for MCMC rungs}
\usage{
optimise_beta(proj, K = 3, target_acceptance = 0.4,
  max_iterations = 100, beta_init = seq(0, 1, l = 10),
  silent = FALSE, ...)
}
\arguments{
\item{proj}{An rgeoprofile project}

\item{K}{The value or values of K to optimise beta for}

\item{target_acceptance}{The minimum acceptance probability to be reached between
MCMC chains before the process terminates}

\item{max_iterations}{Run analyses a maximum number of times before terminating}

\item{beta_init}{An initial set of beta values to run - a sequence from zero
to one}

\item{silent}{print output from console}

\item{...}{extra arguments to be sent to the run_mcmc function}
}
\description{
repeatedly run analysis to find vector of beta values that achieves a required
             coupling acceptance rate between all rungs
}
