% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{overlay_geoprofile}
\alias{overlay_geoprofile}
\title{Add geoprofile to dynamic map}
\usage{
overlay_geoprofile(
  myplot,
  project,
  K = NULL,
  source = NULL,
  realised = FALSE,
  threshold = 0.1,
  col = col_hotcold(),
  opacity = 0.8,
  smoothing = 1,
  legend = FALSE
)
}
\arguments{
\item{myplot}{dynamic map produced by \code{plot_map()} function.}

\item{project}{an RgeoProfile project, as produced by the function
\code{rgeoprofile_project()}.}

\item{K}{which value of K to plot.}

\item{source}{which source to plot. If NULL then plot combined surface.}

\item{realised}{if TRUE then plot surface for realised sources only.}

\item{threshold}{what proportion of geoprofile to plot.}

\item{col}{set of plotting colours.}

\item{opacity}{opacity of geoprofile (that is not invisible due to being
below threshold).}

\item{smoothing}{what level of smoothing to apply to geoprofile. Smoothing is
applied using the \code{raster} function \code{disaggregate}, with
\code{method = "bilinear"}.}

\item{legend}{Set to TRUE or FALSE, this will add a hitscore legend to the
plot.}
}
\description{
Add geoprofile to dynamic map
}
\examples{
\dontshow{p <- rgeoprofile_file("tutorial1_project.rds")}
plot1 <- plot_map()
plot1 <-overlay_geoprofile(myplot = plot1, project = p, K = 2, source = NULL, threshold = 0.5)
plot1
}
