% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{overlay_piecharts}
\alias{overlay_piecharts}
\title{Add pie charts to dynamic map}
\usage{
overlay_piecharts(
  myplot,
  project,
  K = NULL,
  min_size = 10,
  max_size = 30,
  col = NULL
)
}
\arguments{
\item{myplot}{dynamic map produced by \code{plot_map()} function.}

\item{project}{an RgeoProfile project, as produced by the function
\code{rgeoprofile_project()}.}

\item{K}{which value of K to plot.}

\item{min_size, max_size}{lower and upper limits on the size of pie charts.}

\item{col}{segment colours.}
}
\description{
Add pie charts to dynamic map
}
\examples{
\dontshow{p <- rgeoprofile_file("tutorial1_project.rds")}
plot1 <- plot_map()
plot1 <- overlay_piecharts(myplot = plot1, project = p, K = 2)
plot1
}
