% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{overlay_sources}
\alias{overlay_sources}
\title{Add sources to dynamic map}
\usage{
overlay_sources(myplot, lon, lat, icon_url = NULL, icon_width = 20,
  icon_height = 20, icon_anchor_x = 10, icon_anchor_y = 10)
}
\arguments{
\item{myplot}{dynamic map produced by \code{plot_map()} function}

\item{lon}{longitude of sources}

\item{lat}{latitude of sources}

\item{icon_url}{what image to use for the icon}

\item{icon_width}{the width of the icon}

\item{icon_height}{the height of the icon}

\item{icon_anchor_x}{the coordinates of the "tip" of the icon (relative to
its top left corner, i.e. the top left corner means \code{icon_anchor_x =
0} and \code{icon_anchor_y = 0}), and the icon will be aligned so that this
point is at the marker's geographical location}

\item{icon_anchor_y}{the coordinates of the "tip" of the icon (relative to
its top left corner, i.e. the top left corner means \code{icon_anchor_x =
0} and \code{icon_anchor_y = 0}), and the icon will be aligned so that this
point is at the marker's geographical location}
}
\description{
Add sources to dynamic map
}
\examples{
\dontshow{mysim <- rgeoprofile_file("tutorial1_mysim.rds")}
source_locations <- mysim$record$true_source
plot1 <- plot_map()
plot1 <- overlay_sources(myplot = plot1,
                         lon = source_locations$longitude,
                         lat = source_locations$latitude)
plot1
}
