% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{overlay_surface}
\alias{overlay_surface}
\title{Add posterior probability surface to dynamic map}
\usage{
overlay_surface(myplot, project, K = NULL, source = NULL,
  threshold = 0.1, col = rev(col_hotcold()), opacity = 0.8,
  smoothing = 1, legend = FALSE)
}
\arguments{
\item{myplot}{dynamic map produced by \code{plot_map()} function.}

\item{project}{an RgeoProfile project, as produced by the function
\code{rgeoprofile_project()}.}

\item{K}{which value of K to plot.}

\item{source}{which source to plot. If NULL then plot combined surface.}

\item{threshold}{what proportion of posterior probability surface to plot.}

\item{col}{set of plotting colours.}

\item{opacity}{opacity of posterior probability surface (that is not
invisible due to being below threshold).}

\item{smoothing}{what level of smoothing to apply to posterior probability
surface. Smoothing is applied using the \code{raster} function
\code{disaggregate}, with \code{method = "bilinear"}.}

\item{legend}{whether or not a legend is plotted}
}
\description{
Add posterior probability surface to dynamic map
}
\examples{
\dontshow{p <- rgeoprofile_file("tutorial1_project.rds")}
plot1 <- plot_map()
plot1 <-overlay_surface(myplot = plot1, project = p, K = 2, source = NULL, threshold = 0.5)
plot1
}
