% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainFunctions.R
\name{pairwise_distance}
\alias{pairwise_distance}
\title{pairwise_distance}
\usage{
pairwise_distance(points)
}
\arguments{
\item{points}{two sets of x and y points all in latitude and longitude}
}
\description{
Given two sets of points calculate the distance between each pair of points. Returns a matrix whose entries are distances between pairs of points. Also returns the nearest neighbour point in the form of the minimum rowise distance of this matrix.
}
\examples{
some_points <- cbind(1:10, 1:10, 11:20, 11:20)
pairwise_distance(some_points)
}
