% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainFunctions.R
\name{PAjoint}
\alias{PAjoint}
\title{PAjoint}
\usage{
PAjoint(theta_x, theta_y, prior_Sd = NULL, sigma, data, trap_rad, exp_pop)
}
\arguments{
\item{theta_x}{The mean longitude for the source we condition on}

\item{theta_y}{The mean latitude for the source we condition on}

\item{sigma}{The standard deviation of the bivariate normal (source) we condition on}

\item{data}{The trap data consisting of locations in lat/long and associated densities}

\item{trap_rad}{The trap radius we use to approximate the area under the bivariate normal (in kilometres)}

\item{exp_pop}{The expected population over our entire area
'}
}
\description{
Calculate the logged joint probability of observing the data. This acts as the posterior draw. Also extract the logged density of the bivariate normal at each trap location.
}
\examples{
trap_sim <- cbind(runif(10), runif(10), sample(0:5, 10, replace = T))
PAjoint(theta_x = 0, theta_y = 0, prior_Sd = NULL, sigma = 10, data = trap_sim, trap_rad = 2, exp_pop = 25)
}
