% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{PAmcmc}
\alias{PAmcmc}
\title{PAmcmc}
\usage{
PAmcmc(trapData, burnin = 100, iterations = 1000, clusters = 1,
  trap_rad = 1, s_prior_shape = 1, s_prior_rate = 1,
  priorLon = mean(trapData[, 1]), priorLat = mean(trapData[, 2]),
  lon_minMax = range(trapData[, 1]), lat_minMax = range(trapData[, 2]),
  produce_surface = TRUE)
}
\arguments{
\item{trapData}{The main input into the model, consisting of a matrix or data
frame with three columns: trap longitude, latitude and counts}

\item{burnin}{The MCMC's burn in; the number of iterations before the main
sampling begins}

\item{iterations}{The number of iterations in MCMC, this is on top of the
burn in}

\item{clusters}{The number of clusters to be searched for}

\item{trap_rad}{The trap radius in km}

\item{s_prior_shape}{The shape parameter for the prior on the population
density}

\item{s_prior_rate}{The rate parameter for the prior on the population
density}

\item{priorLon}{The prior mean for the source locations (Longitude)}

\item{priorLat}{The prior mean for the source locations (Latitude). Uses mean
of data by default.}

\item{lon_minMax}{The longitudinal limits of the data (already defined)}

\item{lat_minMax}{The latitudinal limits of the data (already defined)}

\item{produce_surface}{Option to produce smoothed surface from posterior
draws}
}
\description{
Run the Metropolis-Hastings algorithm with Gibbs sampler to fit
  the source locations, sigma and or population density
}
\examples{
trap_sim <- cbind(runif(10), runif(10), sample(0:5, 10, replace = TRUE))
PAmcmc(trapData = trap_sim, burnin = 100, iterations = 25e2, clusters = 1, trap_rad = 1,
s_prior_shape = 1, s_prior_rate = 1)
}
