% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainFunctions.R
\name{PAsim}
\alias{PAsim}
\title{PAsim}
\usage{
PAsim(n_true = 1000, sigma_true = 2, K_true = 3, n_traps = 64,
  trap_spacing = "cluster", trapExtent = 3, long_minMax = c(0, 0.5),
  lat_minMax = c(0, 0.5), plotRail = 0.05, trap_const = 0.5,
  trap_clusters = 4, single_count = T, plotting = T, bias = 0)
}
\arguments{
\item{n_true}{The underlying population size}

\item{sigma_true}{The underlying sigma value of the data}

\item{K_true}{The true number of sources}

\item{n_traps}{The number of traps to be used}

\item{trap_spacing}{Set to either "uniform", "random", or "cluster" describes the trap configuration}

\item{trapExtent}{When trap_spacing is clustered, this describes the distance from each trap to the centre of that cluster}

\item{long_minMax}{The longitudinal extent of the data (to be set using this function)}

\item{lat_minMax}{The latitudinal extent of the data (to be set using this function)}

\item{plotRail}{A buffer to be placed around the data when plotting}

\item{trap_const}{This governs the trap radiu, which is the trap_const multiplied by the true sigma}

\item{trap_clusters}{When trap_spacing is clustered, this is sets the number of traps within a cluster}

\item{single_count}{Set to TRUE or FALSE this governs if a n individual can be caught by multiple traps}

\item{plotting}{Set to TRUE or FALSE should the data want to be plotted}

\item{bias}{Set to 0 or 1, if set to 1 two additional sources will be created, one surrounded by empty traps, and one far away from the traps}
}
\description{
A function that simulates and collects all observed and unobserved data
}
\examples{
PAsim(n_true = 100, sigma_true = 1, K_true = 3, n_traps = 10, trap_spacing = "cluster",
trapExtent = 0.5, long_minMax = c(0, 0.1), lat_minMax = c(0, 0.1), plotRail = 0.05,
trap_const = 0.5, trap_clusters = 4, single_count = T, plotting = T, bias = 1)
}
