% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_acf}
\alias{plot_acf}
\title{Produce MCMC autocorrelation plot}
\usage{
plot_acf(project, K = NULL, rung = NULL, col = "black",
  phase = "sampling")
}
\arguments{
\item{project}{an RgeoProfile project, as produced by the function
\code{rgeoprofile_project()}.}

\item{K}{which value of K to plot.}

\item{rung}{which rung to plot. Defaults to the cold chain.}

\item{col}{colour of the trace.}

\item{phase}{plot the acf during the burnin or sampling phase}
}
\description{
Produce MCMC autocorrelation plot of the log-likelihood
}
\examples{
\dontshow{p <- rgeoprofile_file("tutorial1_project.rds")}
plot_acf(project = p)
}
