% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_alpha}
\alias{plot_alpha}
\title{Plot alpha 95\% credible intervals}
\usage{
plot_alpha(project, K = NULL)
}
\arguments{
\item{project}{an RgeoProfile project, as produced by the function
\code{rgeoprofile_project()}.}

\item{K}{which value of K to produce the plot for.}
}
\description{
Plot the over dispersion parameter alpha, in accordance with 
             a negative binomial model
}
\details{
TODO
}
\examples{
# \dontshow{p <- rgeoprofile_file("tutorial1_project.rds")}
# plot_alpha(project = p)
}
