% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_coupling}
\alias{plot_coupling}
\title{Plot acceptance rate between rungs}
\usage{
plot_coupling(project, K = 1, phase = "sampling")
}
\arguments{
\item{project}{an RgeoProfile project, as produced by the function
\code{rgeoprofile_project()}.}

\item{K}{which value of K to produce the plot for.}

\item{phase}{which phase to plot. Must be either "burnin" or "sampling".}
}
\description{
TODO
}
