% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_expected_popsize}
\alias{plot_expected_popsize}
\title{Plot expected population size 95\% credible intervals}
\usage{
plot_expected_popsize(project, K = NULL)
}
\arguments{
\item{project}{an RgeoProfile project, as produced by the function
\code{rgeoprofile_project()}.}

\item{K}{which value of K to produce the plot for.}
}
\description{
Plot credible intervals for the entire expected population (single)
 or individuals sources (independent).
}
\examples{
\dontshow{p <- rgeoprofile_file("tutorial1_project.rds")}
plot_expected_popsize(project = p, K = 1)
}
