% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_loglike_diagnostic}
\alias{plot_loglike_diagnostic}
\title{Produce diagnostic plots of log-likelihood}
\usage{
plot_loglike_diagnostic(project, K = NULL, rung = NULL,
  col = "black")
}
\arguments{
\item{project}{an RgeoProfile project, as produced by the function
\code{rgeoprofile_project()}}

\item{K}{which value of K to plot}

\item{rung}{which rung to plot. Defaults to the cold chain}

\item{col}{colour of the trace}
}
\description{
Produce diagnostic plots of the log-likelihood.
}
\details{
For a value of K produce the auto-correlation, trace and density plots of the MCMC.
}
\examples{
\dontshow{p <- rgeoprofile_file("tutorial1_project.rds")}
plot_loglike_diagnostic(project = p, K = 2)
}
