% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_pseudoAIC}
\alias{plot_pseudoAIC}
\title{Plot the pseudoAIC over all K}
\usage{
plot_pseudoAIC(project)
}
\arguments{
\item{project}{an RgeoProfile project, as produced by the function
\code{rgeoprofile_project()}}
}
\description{
Plot pseudoAIC over all K for the current active parameter set.
}
