% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_trace}
\alias{plot_trace}
\title{Produce MCMC trace plot}
\usage{
plot_trace(project, K = NULL, rung = NULL, col = "black")
}
\arguments{
\item{project}{an RgeoProfile project, as produced by the function
\code{rgeoprofile_project()}}

\item{K}{which value of K to plot}

\item{rung}{which rung to plot. Defaults to the cold chain}

\item{col}{colour of the trace}
}
\description{
Produce MCMC trace plot of the log-likelihood at each iteration.
}
\examples{
\dontshow{p <- rgeoprofile_file("tutorial1_project.rds")}
plot_trace(project = p, K = 2)
Similarly, plot the trace for every K value.
plot_sigma(project = p)
}
