% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{proposal}
\alias{proposal}
\title{Propose parameter values}
\usage{
proposal(n, mu1, mu2, sd)
}
\arguments{
\item{n}{The number of new points to propose}

\item{mu1}{The centre of the bivariate norm to draw from (longitude)}

\item{mu2}{The centre of the bivariate norm to draw from (latitude)}

\item{sd}{The standard deviation of the bivariate norm to draw from
(kilometres)}
}
\description{
Propose new parameter values (specifically a source location)
  from a bivaraite normal.
}
\examples{
proposal(n = 16, mu1 = 0, mu2 = 0, sd = 5)
}
