% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{realised_sources}
\alias{realised_sources}
\title{Calculate realised sources}
\usage{
realised_sources(proj, n_samples = 20, K = 1)
}
\arguments{
\item{proj}{An rgeoprofile project}

\item{n_samples}{how many times we sample from the qmatrix}
}
\description{
Produce a plot that indicates the suitable number of realised 
             sources based upon sampling from the qmatrix.
}
