% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/silverblaze.R
\docType{package}
\name{silverblaze}
\alias{silverblaze}
\alias{silverblaze-package}
\title{Bayesian Geographic Profiling from Presence-Absence Data}
\description{
Bayesian geographic profiling model that allows estimation of
  source locations given presence and absence data.
}
