% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{bvnorm}
\alias{bvnorm}
\title{bvnorm}
\usage{
bvnorm(x, y, mu_x, mu_y, sd, log = TRUE)
}
\arguments{
\item{x}{The location of the point whose density we require (longitude)}

\item{y}{The location of the point whose density we require (latitude)}

\item{mu_x}{The mean of the bivariate normal (longitude)}

\item{mu_y}{The mean of the bivariate normal (latitude)}

\item{sd}{The standard deviation of the bivariate normal (kilometres)}

\item{log}{Whether to return output in log space}
}
\description{
Extract the log density on the bivariate normal at (x, y)
  centred at (mu_x, mu_y)
}
\examples{
bvnorm(x = 1, y = 1, mu_x = 0, mu_y = 0, sd = 10)
}
