% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainFunctions.R
\name{logSum}
\alias{logSum}
\title{logSum}
\usage{
logSum(x)
}
\arguments{
\item{x}{A vector whose values are in log space}
}
\description{
To avoid underflow issues from R, scale the product of small values by xmax.
}
\examples{
x <- log(seq(5e-49, 5e-48, 1e-49))
exp(sum(x))
exp(logSum(x))
}
