% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainFunctions.R
\name{PAmcmc}
\alias{PAmcmc}
\title{PAsim}
\usage{
PAmcmc(trapData, burnin = 100, iterations = 2500, Clusters = 1,
  trap_rad = 1, s_prior_shape = 1, s_prior_rate = 1,
  priorLat = NULL, priorLon = NULL, long_minMax = NULL,
  lat_minMax = NULL)
}
\arguments{
\item{trapData}{The main input into the model consisting of trap longitude, latitude and densities}

\item{burnin}{The MCMC's burn in; the number of iterations before the main sampling begins}

\item{iterations}{The number of iterations in MCMC, this is on top of the burn in}

\item{Clusters}{The number of clusters to be searched for}

\item{trap_rad}{The trap radius defined as some constant times the true sigma value}

\item{s_prior_shape}{The shape parameter for the prior on the population density}

\item{s_prior_rate}{The rate parameter for the prior on the population density}

\item{priorLat}{The prior mean for the source locations (Latitude)}

\item{priorLon}{The prior mean for the source locations (Longitude)}

\item{long_minMax}{The longitudinal limits of the data (already defined)}

\item{lat_minMax}{The latitudinal limits of the data (already defined)}
}
\description{
A function that simulates and collects all observed and unobserved data
}
\examples{
trap_sim <- cbind(runif(10), runif(10), sample(0:5, 10, replace = T))
PAmcmc(trapData = trap_sim, burnin = 100, iterations = 25e2, Clusters = 1, trap_rad = 1,
s_prior_shape = 1, s_prior_rate = 1, priorLat = mean(trap_sim[,2]), priorLon = mean(trap_sim[,1]),
long_minMax = c(min(trap_sim[,1]),max(trap_sim[,1])), lat_minMax = c(min(trap_sim[,2]),max(trap_sim[,2])))
}
